backend/requirements.txt
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-stubs==4.2.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.2
pillow==10.3.0
python-decouple==3.8
ruff==0.4.1
isort==5.13.2